// core/interfaces/page.interface.ts
export enum TopLevelCategory {
  Courses,
  Services,
  Books,
  Products
}

export interface ITopPageAdvantage {
	title: string;
	description: string;
	_id: string;
}

export interface IHhData {
	count: number;
	juniorSalary: number;
	middleSalary: number;
	seniorSalary: number;
	updatedAt: Date;
	_id: string;
}

export interface IBlogData {
	h1: string;
	metaTitle: string;
	metaDescription: string;
	views: number;
	_id: string;
}

export interface ISravnikusData {
	metaTitle: string;
	metaDescription: string;
	qas: any[];
	_id: string;
}

export interface ILearningclubData {
	metaTitle: string;
	metaDescription: string;
	qas: any[];
	_id: string;
}

export interface ITopPageModel {
	_id: string;
	tags: string[];
	secondCategory: string;
	alias: string;
	title: string;
	category: string;
	seoText?: string;
	tagsTitle: string;
	metaTitle: string;
	metaDescription: string;
	firstCategory: TopLevelCategory;
	advantages?: ITopPageAdvantage[];
	createdAt: Date;
	updatedAt: Date;
	__v: number;
	hh?: IHhData;
	qas: any[];
	addresses: any[];
	categoryOn: string;
	blog: IBlogData;
	sravnikus: ISravnikusData;
	learningclub: ILearningclubData;
}